#40 separates components into files
