+++ client/src/components/Groups.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Groups = () => <h2>Your Groups</h2>;
+
+export default Groups;
